LayoutTests/platform/ios-simulator/editing/shadow/select-contenteditable-shadowhost-expected.txt

CONSOLE MESSAGE: line 3: HierarchyRequestError: DOM Exception 3: A Node was inserted somewhere it doesn't belong.
This test confirms that selecting an element having Shadow DOM doesn't cross editing boundaries errornously.

BEFORE
HOST
AFTER
FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
